pyproject.toml
# SPDX-FileCopyrightText: 2025 SensorStim Neurotechnology GmbH <support@capture2go.com>
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "capture2go"
version = "0.8.0"
description = "SDK for Capture2Go, the wearable sensor platform by SensorStim Neurotechnology GmbH."
authors = [
    { name="SensorStim Neurotechnology GmbH", email="support@capture2go.com" }
]
license = "MIT"
license-files = ["LICENSES/*.txt"]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "bleak",
    "pyserial",
]

[project.urls]
"Homepage" = "https://capture2go.com"

[project.optional-dependencies]
docs = [
  "sphinx>=7.0",
  "sphinx-rtd-theme>=2.0.0",
  "myst-parser>=3.0.0",
  "sphinxcontrib-mermaid>=0.9.2",
]

[tool.setuptools.packages.find]
where = ["."]
